Adds Logic to Login Form
In this commit I added the logic necessary to take the users data
and process it in the Flask application.

Firstly @app.route decorator takes two methods; GET and POST,
allowing for the form to be displayed (GET) and the form data to
be submitted (POST).

I added form.validate_on_submit to handle the logical switch
between these scenarios. When the form is requested by GET; the
`if` statement returns False and the flash statement is not
invoked. However the render_template method is returned. When
the form is requested by POST; the `if` statement is True and
the flash method is invoked. The user is then  redirected to
/index.